requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary
pydantic==2.10.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
faker==30.8.2
pillow==11.0.0
python-multipart
alembic==1.12.0
python-dotenv